Ajustes da issue 9
